requirements2.txt
asgiref==3.4.1
boto3==1.9.96
botocore==1.12.96
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==3.2.9
django-crispy-forms==1.13.0
django-heroku==0.3.1
django-storages==1.7.1
docutils==0.14
gunicorn==20.1.0
idna==2.7
jmespath==0.9.3
Pillow==8.4.0
psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2021.3
requests==2.19.1
s3transfer==0.2.0
six==1.12.0
sqlparse==0.4.2
urllib3==1.23
whitenoise==4.1.2
